backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
# using hashlib.pbkdf2_hmac instead of passlib+bcrypt
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
apscheduler==3.10.4
aiosmtplib==3.0.2
httpx==0.28.1
